requirements.txt
langchain 
langchain-openai
datasets
python-dotenv
flask 
langchain_core
langchain_community
langchainhub
chromadb
-e .